Decouple analysis components
